Create SES_1.2.py
Normal legal stuff: don't use this in any way that would break the law... it is for people to secure websites not break them for personal gain...
This is something that will depend on your local area and so before using this script you may want to check local laws to be sure you are not breaking them...
DO NOT BREAK THE LAW!!!
Also should note: this is untested code as for now this is just a hobby to me as of now but posting it here likely help motivate me just because it's here. 
Not to work on this though to be honest it was an interesting time. NOT much more than a POC but if you use the --custom-dorks option with the list
provided at the bottom(or your own), it can have some decent results... but due to the use of 3rd party python lib's(I personally do not want to use this and 
even though it is my code and i love it as a POC: would happily warn anyone against installing all the extras in this to run it). But if you want too
then feel free to enjoy.